Only attempts to remove files if dir not empty

+++ inlineReactToCSS/reset.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
-rm ./input/*
+[[ $(ls -A ./input) ]] && rm ./input/*
 cp ./inline/* ./input
 prettier --write ./input/*
